requirements.txt
streamlit==1.8.1
torch==1.10.0
torchvision==0.11.1
Pillow==7.1.2
matplotlib==3.2.2
